fix(webui): share remote address selection
