+++ src/process/webserver/index.ts
@@ -9,7 +9,6 @@ import net from 'net';
 import { createServer } from 'http';
 import { WebSocketServer } from 'ws';
 import { execSync } from 'child_process';
-import { networkInterfaces } from 'os';
 import { AuthService } from '@process/webserver/auth/service/AuthService';
 import { UserRepository } from '@process/webserver/auth/repository/UserRepository';
 import { AUTH_CONFIG, SERVER_CONFIG } from './config/constants';
@@ -19,6 +18,7 @@ import { registerAuthRoutes } from './routes/authRoutes';
 import { registerApiRoutes } from './routes/apiRoutes';
 import { registerStaticRoutes, resolveRendererPath, VITE_DEV_PORT } from './routes/staticRoutes';
 import { generateQRLoginUrlDirect } from '@process/bridge/webuiQR';
+import { getPreferredRemoteAddress } from './networkAddress';
 
 // Express Request 类型扩展定义在 src/webserver/types/express.d.ts
 // Express Request type extension is defined in src/webserver/types/express.d.ts
@@ -56,29 +56,6 @@ export function getInitialAdminPassword(): string | null {
  */
 export function clearInitialAdminPassword(): void {
   initialAdminPassword = null;
-}
-
-/**
- * 获取局域网 IP 地址
- * Get LAN IP address using os.networkInterfaces()
- */
-function getLanIP(): string | null {
-  const nets = networkInterfaces();
-  for (const name of Object.keys(nets)) {
-    const netInfo = nets[name];
-    if (!netInfo) continue;
-
-    for (const iface of netInfo) {
-      // 跳过内部地址（127.0.0.1）和 IPv6
-      // Skip internal addresses (127.0.0.1) and IPv6
-      const isIPv4 = iface.family === 'IPv4';
-      const isNotInternal = !iface.internal;
-      if (isIPv4 && isNotInternal) {
-        return iface.address;
-      }
-    }
-  }
-  return null;
 }
 
 /**
@@ -127,7 +104,7 @@ function getServerIP(): string | null {
 
   // 2. 所有平台：获取局域网 IP（包括 Windows/Mac/Linux）
   // All platforms: get LAN IP (Windows/Mac/Linux)
-  return getLanIP();
+  return getPreferredRemoteAddress();
 }
 
 /**
